Redesign of collocations
